requirements.txt
Flask
Flask-login
Flask-wtf
cassandra-driver
passlib
